Refactor. Renamed all Bucketers to Dimensions
