+++ examples/policies/ITwitter-ORedis.json
@@ -22,19 +22,19 @@
   ],
   "dimensions": [
     {
-      "dimensionType": "TwitterStatusBucketer",
+      "dimensionType": "TwitterStatusDimension",
       "name": "status"
     },
     {
-      "dimensionType": "PassthroughBucketer",
+      "dimensionType": "PassthroughDimension",
       "name": "wordsN"
     },
     {
-      "dimensionType": "DateTimeBucketer",
+      "dimensionType": "DateTimeDimension",
       "name": "timestamp"
     },
     {
-      "dimensionType": "GeoHashBucketer",
+      "dimensionType": "GeoHashDimension",
       "name": "geolocation"
     }
   ],
