+++ plugins/dimension-geoHash/.gitignore
